requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.10.11=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=5.0.1=pypi_0
attrdict=2.0.1=pypi_0
attrs=24.2.0=pypi_0
bezier=2020.5.19=pypi_0
ca-certificates=2024.8.30=hbcca054_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.8.2=pypi_0
conda-pack=0.7.1=py38h06a4308_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
deepdiff=8.0.1=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt-ng=0.9.0=pypi_0
enum34=1.1.10=pypi_0
evdev=1.7.1=pypi_0
exceptiongroup=1.2.2=pypi_0
farama-notifications=0.0.4=pypi_0
faster-fifo=1.4.7=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
future=1.0.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.35.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.66.2=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
gymnasium=0.29.1=pypi_0
huggingface-hub=0.25.2=pypi_0
idna=3.10=pypi_0
imageio=2.35.1=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
iniconfig=2.0.0=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonpickle=3.3.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
keyboard=0.13.5=pypi_0
kiwisolver=1.4.7=pypi_0
lbforaging=2.0.0=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.41.1=pypi_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matrixgames=0.0.2=pypi_0
mock=5.1.0=pypi_0
mpmath=1.3.0=pypi_0
mpyq=0.2.5=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
munch=4.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
noise=1.2.2=pypi_0
numba=0.58.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.77=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.15=h5eee18b_0
orderly-set=5.2.2=pypi_0
packaging=24.1=pypi_0
pandas=2.0.1=pypi_0
pbr=6.1.0=pypi_0
pettingzoo=1.24.3=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.3.6=pypi_0
plotly=5.13.0=pypi_0
pluggy=1.5.0=pypi_0
portpicker=1.6.0=pypi_0
probscale=0.2.5=pypi_0
propcache=0.2.0=pypi_0
protobuf=3.20.3=pypi_0
psutil=6.0.0=pypi_0
py=1.11.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pygame=2.6.1=pypi_0
pyglet=1.5.23=pypi_0
pynput=1.7.7=pypi_0
pyopengl=3.1.7=pypi_0
pyparsing=3.1.4=pypi_0
pysc2=3.0.0=pypi_0
pytest=8.3.3=pypi_0
python=3.8.20=he870216_0
python-dateutil=2.9.0.post0=pypi_0
python-xlib=0.33=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
ray=2.10.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rpds-py=0.20.1=pypi_0
rsa=4.9=pypi_0
rtree=1.0.0=pypi_0
rware=2.0.0=pypi_0
s2clientprotocol=4.10.1.75800.0=pypi_0
s2protocol=5.0.13.92440.4=pypi_0
sacred=0.8.6=pypi_0
sample-factory=2.1.1=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
sentry-sdk=2.16.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.1.0=py38h06a4308_0
shimmy=2.0.0=pypi_0
signal-slot-mp=1.0.5=pypi_0
six=1.16.0=pypi_0
sk-video=1.1.10=pypi_0
smac=1.0.0=pypi_0
smaclite=0.0.1=pypi_0
smacv2=1.0.0=pypi_0
smmap=5.0.1=pypi_0
snakeviz=2.2.0=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboard-logger=0.1.0=pypi_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.0.2=pypi_0
torch=2.4.1=pypi_0
torch-geometric=2.6.1=pypi_0
torch-scatter=2.1.2=pypi_0
torchvision=0.19.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.5=pypi_0
transforms3d=0.4.2=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
vmas=1.4.3=pypi_0
wandb=0.18.3=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.44.0=py38h06a4308_0
whichcraft=0.6.1=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.15.2=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
